+++ packages/storybook/src/web-components/alert/alert.mdx
@@ -40,16 +40,6 @@ import { IncludeDocumentation } from "../../utils/IncludeDocumentation";
 
 ## Inhoud
 
-### Custom Icon
-
-<Description of={AlertStories.Icon} />
-<Canvas of={AlertStories.Icon} />
-
-### No Icon
-
-<Description of={AlertStories.NoIcon} />
-<Canvas of={AlertStories.NoIcon} />
-
 ### Aria `role=alert`
 
 <Description of={AlertStories.AriaRole} />
