+++ packages/storybook/src/web-components/alert/alert.stories.tsx
@@ -2,7 +2,6 @@ import tokens from '@lux-design-system/design-tokens/dist/index.json';
 import {
   LuxAlert as Alert,
   LuxButton as Button,
-  LuxIconMail as IconMail,
   LuxParagraph as Paragraph,
 } from '@lux-design-system/web-components-react';
 import type { JSX } from '@lux-design-system/web-components-stencil';
@@ -10,6 +9,7 @@ import { useState } from '@storybook/preview-api';
 import type { Meta, StoryObj } from '@storybook/react';
 import { type HTMLAttributes, type PropsWithChildren } from 'react';
 import tokensDefinition from './tokens.json';
+import { VisualTypes } from './visual/Types';
 import { createDesignTokensStory, createVisualRegressionStory, VisualRegressionWrapper } from '../../utils';
 
 const LuxAlert = (props: PropsWithChildren<JSX.LuxAlert> & HTMLAttributes<HTMLLuxButtonElement>) => (
@@ -142,37 +142,6 @@ export const Error: Story = {
   ),
 };*/
 
-export const Icon: Story = {
-  name: 'Custom Icon',
-  args: {
-    type: 'info',
-    children: textTemplate('Alert with custom icon'),
-  },
-  parameters: {
-    docs: {
-      description: {
-        story: 'Een custom icon kan geplaats worden via `slot="icon"`.',
-      },
-    },
-  },
-  render: (args) => (
-    <LuxAlert {...args}>
-      <LuxParagraph>{args.children}</LuxParagraph>
-      <IconMail slot="icon" />
-    </LuxAlert>
-  ),
-};
-
-export const NoIcon: Story = {
-  ...AlertTemplate,
-  name: 'No Icon',
-  args: {
-    type: 'info',
-    hideIcon: true,
-    children: textTemplate('No Icon alert'),
-  },
-};
-
 export const AriaRole: Story = {
   ...AlertTemplate,
   name: 'Aria Role Alert',
@@ -213,27 +182,15 @@ export const AriaRole: Story = {
 
 export const DesignTokens = createDesignTokensStory(meta);
 
-const VisualAlertTemplate = ({ children, type }: PropsWithChildren<JSX.LuxAlert>) => (
-  <LuxAlert type={type}>
-    <LuxParagraph>{children}</LuxParagraph>
-  </LuxAlert>
-);
-
 export const Visual = createVisualRegressionStory(() => (
   <>
     <h4 className="lux-heading-3">Light</h4>
     <VisualRegressionWrapper className={`lux-theme--logius-light`}>
-      <VisualAlertTemplate type="info">Info Alert</VisualAlertTemplate>
-      <VisualAlertTemplate type="ok">Okay Alert</VisualAlertTemplate>
-      <VisualAlertTemplate type="warning">Warning Alert</VisualAlertTemplate>
-      <VisualAlertTemplate type="error">Error Alert</VisualAlertTemplate>
+      <VisualTypes />
     </VisualRegressionWrapper>
     <h4 className="lux-heading-3">Dark</h4>
     <VisualRegressionWrapper className={`lux-theme--logius-dark`}>
-      <VisualAlertTemplate type="info">Info Alert</VisualAlertTemplate>
-      <VisualAlertTemplate type="ok">Okay Alert</VisualAlertTemplate>
-      <VisualAlertTemplate type="warning">Warning Alert</VisualAlertTemplate>
-      <VisualAlertTemplate type="error">Error Alert</VisualAlertTemplate>
+      <VisualTypes />
     </VisualRegressionWrapper>
   </>
 ));
